+++ Components/WAInputDateTime.razor.cs
@@ -157,11 +157,13 @@ namespace WebAwesomeBlazor.Components
             {
                 return Size switch
                 {
-                    InputSize.Small => "small",
-                    InputSize.Medium => "medium",
-                    InputSize.Large => "large",
+                    InputSize.XSmall => "xs",
+                    InputSize.Small => "s",
+                    InputSize.Medium => "m",
+                    InputSize.Large => "l",
+                    InputSize.XLarge => "xl",
                     InputSize.Inherit => "inherit",
-                    _ => "inherit"
+                    _ => "m"
                 };
             }
         }
@@ -187,7 +189,7 @@ namespace WebAwesomeBlazor.Components
 
                 formattedValue = GetFormattedValue(Value!);
                 await LoadModuleAsync("./_content/WebAwesomeBlazor/Components/WAInput.razor.js");
-                await SafeInvokeVoidAsync("initialize", Id!, objRef, Value!);
+                _instance = await SafeInvokeAsync<IJSObjectReference>("initialize", Id!, objRef, Value!);
 
                 //await ValueChanged.InvokeAsync(Value);
             }
@@ -230,6 +232,25 @@ namespace WebAwesomeBlazor.Components
                 await SetValueAsync(oldValue!, Value);
                 oldValue = Value;
             }
+        }
+
+        protected override async ValueTask DisposeAsyncCore(bool disposing)
+        {
+            if (disposing)
+            {
+                try
+                {
+                    if (_instance is not null)
+                        await _instance.InvokeVoidAsync("dispose");
+                }
+                catch (JSDisconnectedException)
+                {
+                }
+
+                objRef?.Dispose();
+
+            }
+
         }
         #endregion