+++ Components/WATag.razor.cs
@@ -90,11 +90,13 @@ namespace WebAwesomeBlazor.Components
             {
                 return Size switch
                 {
-                    TagSize.Small => "small",
-                    TagSize.Medium => "medium",
-                    TagSize.Large => "large",
+                    TagSize.XSmall => "xs",
+                    TagSize.Small => "s",
+                    TagSize.Medium => "m",
+                    TagSize.Large => "l",
+                    TagSize.XLarge => "xl",
                     TagSize.Inherit => "inherit",
-                    _ => "inherit"
+                    _ => "m"
                 };
             }
         }
@@ -107,23 +109,37 @@ namespace WebAwesomeBlazor.Components
         #region Lifecycle
         protected override async Task OnAfterRenderAsync(bool firstRender)
         {
+            await base.OnAfterRenderAsync(firstRender);
             if (firstRender)
             {
-                objRef = DotNetObjectReference.Create(this);
-                await SafeInvokeVoidAsync("initialize", Id!, objRef);
+
+                _instance = await SafeInvokeAsync<IJSObjectReference>("initialize", Id!, objRef);
             }
         }
-
 
         protected override async ValueTask DisposeAsyncCore(bool disposing)
         {
             if (disposing)
             {
+                try
+                {
+                    if (_instance is not null)
+                        await _instance.InvokeVoidAsync("dispose");
 
+
+                }
+                catch (JSDisconnectedException)
+                {
+                }
                 objRef?.Dispose();
             }
 
-            await base.DisposeAsyncCore(disposing);
+        }
+
+        protected override async Task OnInitializedAsync()
+        {
+            objRef ??= DotNetObjectReference.Create(this);
+            await base.OnInitializedAsync();
         }
         #endregion